refactor: refactor bridge service tests
Refactored the bridge service tests to improve code organization and readability.

Changes:
- Removed unused imports and unnecessary struct definitions.
- Reorganized import statements for better clarity.
- Updated references to blockchain and bridge service modules.
- Removed redundant implementations and conversions.
- Refactored monitoring structs related to bridge contracts.
